extracting public pipeline APIs into PipelineAPISubservice
